website/requirements.txt
flask
flask-login
flask-sqlalchemy
Jinja2
paho-mqtt

